CHANGED: pass image buffer to jimp, implement try-catch when performing dir read

+++ src/helpers/imageHelpers.js
@@ -3,18 +3,15 @@ import Jimp from 'jimp';
 import path from 'path';
 import fs from 'fs';
 
-const extractTextFromImage = async () => {
+const extractTextFromImage = async (imageBuffer) => {
   const worker = createWorker({
     cachePath: path.join(__dirname, '..', 'lang-data'),
   });
 
   let rectangle;
 
-  Jimp.read(
-    path.join(__dirname, '..', 'test-images', 'original.png'),
-    (err, image) => {
-      if (err) throw err;
-      // get our image size and update our rectangle for parsing only part of the image
+  const success = Jimp.read(imageBuffer)
+    .then((image) => {
       const { height, width } = image.bitmap;
       rectangle = {
         left: 0,
@@ -28,32 +25,39 @@ const extractTextFromImage = async () => {
         .invert() // invert colors
         .contrast(1) // boost contrast
         .write(path.join(__dirname, '..', 'test-images', 'modified.png')); // save
+      return true;
+    })
+    .catch(() => {
+      return false;
+    });
+
+  if (success) {
+    await worker.load();
+    await worker.loadLanguage('eng');
+    await worker.initialize('eng');
+    const {
+      data: { text },
+    } = await worker.recognize(
+      path.join(__dirname, '..', 'test-images', 'modified.png'),
+      { rectangle }
+    );
+
+    await worker.terminate();
+
+    const files = fs.readdirSync(path.join(__dirname, '..', 'test-images'));
+
+    if (Array.isArray(files)) {
+      try {
+        files.forEach((file) =>
+          fs.unlinkSync(path.join(__dirname, '..', 'test-images', file))
+        );
+      } catch {
+        return false;
+      }
     }
-  );
-
-  await worker.load();
-  await worker.loadLanguage('eng');
-  await worker.initialize('eng');
-  const {
-    data: { text },
-  } = await worker.recognize(
-    path.join(__dirname, '..', 'test-images', 'modified.png'),
-    { rectangle }
-  );
-
-  await worker.terminate();
-
-  fs.readdir(path.join(__dirname, '..', 'test-images'), (err, files) => {
-    if (err) throw err;
-
-    if (files.length) {
-      files.forEach((file) =>
-        fs.unlinkSync(path.join(__dirname, '..', 'test-images', file))
-      );
-    }
-  });
-
-  return text;
+    return text;
+  }
+  return false;
 };
 
 export default extractTextFromImage;
